old and new bugs fixed in MMG5_split3op()
In the case this function is generating 5 tets, the previous commit left isxt[4]
undefined. Before that, isxt[4] was always set to zero. The full Mmg test
battery was succesful on both, but my "david" cases failed spectacularly on the
previous commit, with coquilFace errors and such. Now all isxt members are
properly computed.